python/pynopegl-utils/requirements.txt
pillow
pynopegl
pyside6
watchdog
numpy
jinja2
